+++ src/DownloadDeps.py
@@ -387,14 +387,6 @@ class DownloadDependencies:
                 return_codes.append(return_code)
 
                 if backend == "tensorrt":
-                    if torch_version == "2.7.0":
-                        torch_trt_link = "https://github.com/TNTwise/real-video-enhancer-models/releases/download/models/torch_tensorrt-2.7.0-cp312-cp312-"
-                        if PLATFORM == 'linux':
-                            torch_trt_link += "linux_x86_64.whl"
-                        elif PLATFORM == 'win32':
-                            torch_trt_link += "win_amd64.whl"
-                    else:
-                        torch_trt_link = "torch-tensorrt==2.6.0+cu126"
                     
                     deps = [
                         "tensorrt==10.9.0.34",
@@ -404,7 +396,7 @@ class DownloadDependencies:
                         
                     ]
                     if install:
-                        deps += ["--no-deps",f"torch-tensorrt=={torch_version}"]
+                        deps += ["--no-deps",f"torch-tensorrt=={torch_version}+{torch_backend}"]
 
                     return_code = self.pip(deps, install)
                     return_codes.append(return_code)